requirements.txt
matplotlib==3.9.2
numpy==2.1.3
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
PyYAML==6.0.2
scipy==1.14.1
tensorboard==2.18.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.6
rich==14.1.0
discord-webhook==1.4.1
timm==1.0.20
seaborn==0.13.2
openpyxl==3.1.5